+++ NEWS.md
@@ -3,7 +3,7 @@
 ## MsCoreUtils 1.11.1
 
 - Add a `margin` argument to (relevant) imputation functions to
-  support (and make it explicit) along with dimensions (row or
+  support (and make it explicit) along which dimensions (row or
   columns) imputation is performed.
 
 ## MsCoreUtils 1.11.0